refactor(module:tag): refactor
