--- components/tag/nz-tag.component.html
@@ -1,8 +1,0 @@
-<ng-content></ng-content>
-<i
-  nz-icon
-  nzType="close"
-  *ngIf="nzMode === 'closeable'"
-  tabindex="-1"
-  (click)="closeTag($event)"
-></i>
