+++ components/tag/public-api.ts
@@ -6,5 +6,5 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-export * from './nz-tag.component';
-export * from './nz-tag.module';
+export * from './tag.component';
+export * from './tag.module';
