+++ components/tag/tag.module.ts
@@ -11,7 +11,7 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 
-import { NzTagComponent } from './nz-tag.component';
+import { NzTagComponent } from './tag.component';
 
 @NgModule({
   imports: [CommonModule, FormsModule, NzIconModule],
